+++ backend/src/main/java/cn/zimu/fulfillment/connector/SourceShipmentResult.java
@@ -1,20 +1,22 @@
 package cn.zimu.fulfillment.connector;
 
 import cn.zimu.fulfillment.common.domain.SourceChannel;
-import java.math.BigDecimal;
 
 /**
  * 回传发货结果（ticket 01，Phase 2 契约，签名见设计文档 §4.1）。
  *
  * <p>输入只用内部标准字段：来源引用、来源行引用、实际发货数量、履约结果、
  * 来源渠道承运商输出值、首批运单号与异常原因；不得把平台表格列名泄露到领域层。</p>
+ *
+ * <p>{@code sourceUnitQuantity} 仅在本地数量无法精确换算、结果只用于展示阻断检查时可为
+ * {@code null}；任何 Connector 写入路径都必须先拒绝该空值，不能伪造 0 份。</p>
  */
 public record SourceShipmentResult(
         SourceChannel channel,
         String sourceRef,
         String sourceLineRef,
-        BigDecimal actualShippedQuantity,
-        BigDecimal sourceUnitQuantity,
+        long actualShippedQuantity,
+        Long sourceUnitQuantity,
         String outcome,
         String carrierOutputValue,
         String firstTrackingNo,
@@ -30,8 +32,8 @@ public record SourceShipmentResult(
             SourceChannel channel,
             String sourceRef,
             String sourceLineRef,
-            BigDecimal actualShippedQuantity,
-            BigDecimal sourceUnitQuantity,
+            long actualShippedQuantity,
+            Long sourceUnitQuantity,
             String outcome,
             String carrierOutputValue,
             String firstTrackingNo,
@@ -51,7 +53,7 @@ public record SourceShipmentResult(
             SourceChannel channel,
             String sourceRef,
             String sourceLineRef,
-            BigDecimal actualShippedQuantity,
+            long actualShippedQuantity,
             String outcome,
             String carrierOutputValue,
             String firstTrackingNo,
@@ -71,7 +73,7 @@ public record SourceShipmentResult(
             SourceChannel channel,
             String sourceRef,
             String sourceLineRef,
-            BigDecimal actualShippedQuantity,
+            long actualShippedQuantity,
             String outcome,
             String carrierOutputValue,
             String firstTrackingNo,
